+++ NEWS.md
@@ -6,6 +6,10 @@ matrix, numeric data.frame, or a vector denoting the mutation **counts**.
 * Added new internal function `AdjustNumberOfCores` to change the number of cores
 automatically to 1 if the operating system is Windows.
 
+## Changed
+* Updated plotting function for SBS96Catalog to adjust the sample name position
+when the catalog type is counts.
+
 <br/>
 
 # ICAMS 2.2.6